Refactor - extract a generalized composite service filter
