+++ src/Filters/CSS/Composite/Filter.php
@@ -4,20 +4,16 @@
 namespace Kibo\Phast\Filters\CSS\Composite;
 
 use Kibo\Phast\Filters\Service\CachedResultServiceFilter;
-use Kibo\Phast\Logging\LoggingTrait;
-use Kibo\Phast\Services\ServiceFilter;
+use Kibo\Phast\Filters\Service\CompositeFilter;
 use Kibo\Phast\ValueObjects\Resource;
 use Kibo\Phast\Filters\CSS\CommentsRemoval;
 
-class Filter implements CachedResultServiceFilter {
-    use LoggingTrait;
+class Filter extends CompositeFilter implements CachedResultServiceFilter {
 
     /**
      * @var string
      */
     protected $serviceUrl;
-
-    protected $filters = [];
 
     /**
      * Filter constructor.
@@ -26,10 +22,6 @@ class Filter implements CachedResultServiceFilter {
     public function __construct($serviceUrl) {
         $this->serviceUrl = $serviceUrl;
         $this->filters[] = new CommentsRemoval\Filter();
-    }
-
-    public function addFilter(ServiceFilter $filter) {
-        $this->filters[] = $filter;
     }
 
     public function getCacheHash(Resource $resource, array $request) {
@@ -42,19 +34,4 @@ class Filter implements CachedResultServiceFilter {
         }
         return join("\n", $parts);
     }
-
-    public function apply(Resource $resource, array $request) {
-        $this->logger()->info('Starting filtering for resource {url}', ['url' => $resource->getUrl()]);
-        $result = array_reduce(
-            $this->filters,
-            function (Resource $resource, ServiceFilter $filter) use ($request) {
-                $this->logger()->info('Starting {filter}', ['filter' => get_class($filter)]);
-                return $filter->apply($resource, $request);
-            },
-            $resource
-        );
-        $this->logger()->info('Done filtering for resource {url}', ['url' => $resource->getUrl()]);
-        return $result;
-    }
-
 }
