+++ test/Filters/CSS/Composite/FilterTest.php
@@ -20,34 +20,6 @@ class FilterTest extends TestCase {
         $this->filter = new Filter('service-url');
     }
 
-    public function testReturnSameResourceWhenEmpty() {
-        $resource = $this->makeResource();
-        $returned = $this->filter->apply($resource, []);
-        $this->assertEquals($resource->getContent(), $returned->getContent());
-    }
-
-    public function testApplyingFilters() {
-        $resource0 = $this->makeResource();
-        $resource1 = $this->makeResource();
-        $resource2 = $this->makeResource();
-        $filter1 = $this->createMock(ServiceFilter::class);
-        $filter1->expects($this->once())
-            ->method('apply')
-            ->with($resource0, [])
-            ->willReturn($resource1, []);
-        $filter2 = $this->createMock(ServiceFilter::class);
-        $filter2->expects($this->once())
-            ->method('apply')
-            ->with($resource1, [])
-            ->willReturn($resource2);
-
-        $this->filter->addFilter($filter1);
-        $this->filter->addFilter($filter2);
-        $returned = $this->filter->apply($resource0, []);
-
-        $this->assertEquals($resource2->getContent(), $returned->getContent());
-    }
-
     public function testGetCacheHash() {
         $resource = Resource::makeWithContent(URL::fromString('http://phast.test'), 'the-content');
 
@@ -80,10 +52,6 @@ class FilterTest extends TestCase {
         $resource = Resource::makeWithContent(URL::fromString('http://phast.test'), $css);
         $filtered = $this->filter->apply($resource, []);
         $this->assertEquals(' selector {rule:  value}', $filtered->getContent());
-    }
-
-    private function makeResource() {
-        return Resource::makeWithContent(URL::fromString('http://phast.test'), 'content');
     }
 
 }
